clean the directory(change name, rm...)
