+++ myTrie.hpp
@@ -1,3 +1,8 @@
+#include <stddef.h>
+#include <cstdint>
+#include <cstdlib>
+#include <vector>
+
 #include <iostream>
 #include <sstream>
 #include <stack>
@@ -6,41 +11,47 @@
 #include <fstream>
 #include <ios>
 #include <iostream>
+#include <set>
 #include <string>
 
-// #define BUCKET_INIT_COUNT 32
-// #define BURST_POINT 16384
+#include <assert.h>
 
-#define BUCKET_INIT_COUNT 5
-#define BURST_POINT 80
+#define DEFAULT_Associativity 8
+#define DEFAULT_Bucket_num 10
+#define DEFAULT_Max_bytes_per_kv 1000
+#define DEFAULT_Burst_ratio 0.5
+
+// configuration
+static size_t Associativity;
+static size_t Bucket_num;
+static size_t Max_bytes_per_kv;
+static double Burst_ratio;
+static size_t Max_slot_num;
+static size_t Max_loop;
+
+using namespace std;
 
 namespace myTrie {
 // charT = char, T = value type, keysizeT = the type describe keysize
 template <class CharT, class T, class KeySizeT = std::uint16_t>
 class htrie_map {
    public:
-    // burst_threshold should be set to be greater than 26 for 26 alaphbet
-    // and ", @ .etc.(the test in lubm40 shows that it has 50 char species)
-    static const size_t DEFAULT_BURST_THRESHOLD = BURST_POINT;
-    static const size_t DEFAULT_BUCKET_INIT_COUNT = BUCKET_INIT_COUNT;
-    size_t burst_threshold;
-    size_t bucket_num;
-
+    // DEFAULT_Associativity * DEFAULT_Bucket_num should be set to be greater
+    // than 26 for 26 alaphbet and ", @ .etc.(the test in lubm40 shows that it
+    // has 50 char species)
     enum class node_type : unsigned char { HASH_NODE, TRIE_NODE };
     class trie_node;
     class hash_node;
-    class array_bucket;
+    class iterator;
 
     class anode {
        public:
         node_type _node_type;
         trie_node* parent;
-        CharT myChar;
 
         bool isHashNode() { return _node_type == node_type::HASH_NODE; }
         bool isTrieNode() { return _node_type == node_type::TRIE_NODE; }
         void setParent(trie_node* p) { parent = p; }
-        void setChar(CharT c) { myChar = c; }
 
         void deleteMe() {
             if (this->isTrieNode()) {
@@ -57,13 +68,16 @@ class htrie_map {
     class trie_node : public anode {
        public:
         // store the suffix of hash_node or trie_node
-        std::map<CharT, anode*> childs;
+        std::map<CharT, trie_node*> childs;
         hash_node* onlyHashNode;
+        CharT myChar;
 
         trie_node(CharT c, trie_node* p) {
             anode::_node_type = node_type::TRIE_NODE;
-            anode::myChar = c;
             anode::parent = p;
+
+            myChar = c;
+            onlyHashNode = nullptr;
         }
 
         ~trie_node() {
@@ -72,26 +86,22 @@ class htrie_map {
             free(onlyHashNode);
         }
 
-        trie_node* create_trienode_With(CharT key,
-                                        size_t customized_burst_threshold,
-                                        size_t customized_bucket_count) {
-            trie_node* new_trie_node = new trie_node(key, this);
-            new_trie_node->onlyHashNode = new hash_node(
-                customized_burst_threshold, customized_bucket_count);
-            new_trie_node->onlyHashNode->anode::parent = new_trie_node;
-            this->addChildTrieNode(new_trie_node);
-            return new_trie_node;
-        }
-
-        anode* findChildNode(CharT c) {
+        anode* findChildNode(CharT c, bool findMode) {
             auto found = childs.find(c);
             if (found != childs.end()) {
+                trie_node* target = found->second;
+                if (target->onlyHashNode != nullptr)
+                    return target->onlyHashNode;
                 return found->second;
             } else {
-                if (childs.size() == 0) {
-                    return onlyHashNode;
+                if (findMode) {
+                    return nullptr;
+                } else {
+                    trie_node* son_trie_node = new trie_node(c, this);
+                    this->addChildTrieNode(son_trie_node);
+                    son_trie_node->onlyHashNode = new hash_node(son_trie_node);
+                    return son_trie_node->onlyHashNode;
                 }
-                return nullptr;
             }
         }
 
@@ -100,7 +110,7 @@ class htrie_map {
         hash_node* getOnlyHashNode() { return onlyHashNode; }
 
         void addChildTrieNode(trie_node* node) {
-            childs[node->anode::myChar] = node;
+            childs[node->myChar] = node;
             // clear the onlyHashNode because a trie_node will only have childs
             // or have the onlyHashNode
             onlyHashNode = nullptr;
@@ -109,367 +119,533 @@ class htrie_map {
 
     class hash_node : public anode {
        public:
-        std::vector<array_bucket> kvs;
+        class slot;
+
+        slot* key_metas;
+        vector<std::pair<char*, size_t>> pages;
+        size_t elem_num;
+        size_t cur_page_id;
+
         T onlyValue;
         bool haveValue;
-        size_t hn_burst_threshold;
-        size_t hn_bucket_num;
-        uint32_t element_count;
 
-        hash_node(size_t customized_burst_threshold,
-                  size_t customized_bucket_count) {
+       public:
+        struct slot {
+            KeySizeT length;
+            size_t pos;
+            size_t page_id;
+
+            bool isEmpty() { return length == 0; }
+
+            slot(KeySizeT l, size_t p, size_t pi)
+                : length(l), pos(p), page_id(pi) {}
+
+            void set_slot(KeySizeT l, size_t p, size_t pi) {
+                length = l;
+                pos = p;
+                page_id = pi;
+            }
+        };
+
+       public:
+        explicit hash_node(trie_node* p)
+            : elem_num(0), cur_page_id(0), onlyValue(T()), haveValue(false) {
             anode::_node_type = node_type::HASH_NODE;
-            anode::parent = nullptr;
+            anode::parent = p;
+            key_metas = (slot*)malloc(Max_slot_num * sizeof(slot));
 
-            hn_burst_threshold = customized_burst_threshold;
-            hn_bucket_num = customized_bucket_count;
-
-            kvs = std::vector<array_bucket>(hn_bucket_num);
-
-            onlyValue = T();
-            haveValue = false;
-
-            element_count = 0;
-        }
-
-        ~hash_node() {
-            std::vector<array_bucket> empty;
-            for (auto it = kvs.begin(); it != kvs.end(); it++) {
-                free(it->arr_buffer);
+            // init key space
+            for (int i = 0; i != Max_slot_num; i++) {
+                key_metas[i].length = 0;
+                key_metas[i].pos = 0;
+                key_metas[i].page_id = 0;
             }
-            kvs.swap(empty);
+
+            char* page = (char*)malloc(Max_bytes_per_kv);
+            pages.push_back(std::pair<char*, size_t>(page, 0));
         }
 
-        bool need_burst() { return element_count > hn_burst_threshold; }
+        // todo
+        ~hash_node() {}
 
-        T& access_onlyValue_in_hashnode(
-            htrie_map* hm,
-            typename myTrie::htrie_map<CharT, T>::SearchPoint* sp) {
-            haveValue = true;
-            if (sp != nullptr) {
-                sp->hnode = this;
-                sp->abucket = nullptr;
-                sp->pos = 0;
+        inline char* get_tail_pointer(slot* s) {
+            return pages[s->page_id].first + s->pos;
+        }
+
+        std::pair<bool, T> find_kv_in_pages(slot& s, const CharT* key,
+                                            size_t keysize) const {
+            if (myTrie::hashRelative::keyEqual(pages[s.page_id].first + s.pos,
+                                               s.length, key, keysize)) {
+                return std::pair<bool, T>(
+                    true, *((T*)(pages[s.page_id].first + s.pos + s.length)));
             }
-            return onlyValue;
+            return std::pair<bool, T>(false, T());
         }
 
-        // find the key with keysize in its kvs
-        T& access_kv_in_hashnode(
-            const CharT* key, size_t keysize, htrie_map* hm,
-            typename myTrie::htrie_map<CharT, T>::SearchPoint* sp) {
-            hash_node* targetNode = this;
-            size_t move_pos = 0;
-
-            // burst if hash_node have too many element
-            if (need_burst()) {
-                std::map<std::string, T> elements;
-                // add new or existed key, if existed, it will be overwritten at
-                // the loop below
-                elements[std::string(key, keysize)] = T{};
-
-                // gather elements in 'this' hash_node to burst
-                for (auto it = kvs.begin(); it != kvs.end(); it++) {
-                    std::vector<std::pair<std::string, T>> bucket_element;
-                    it->get_item_in_array_bucket(bucket_element);
-                    for (auto itt = bucket_element.begin();
-                         itt != bucket_element.end(); itt++) {
-                        elements[itt->first] = itt->second;
-                    }
+        // return <found?, iterator>
+        // iterator:    if slotid==-1, bucket is full
+        //              if slotid!=-1, slotid is the insert position
+        std::pair<bool, iterator> find_in_bucket(size_t bucketid,
+                                                 const CharT* key,
+                                                 size_t keysize) {
+            slot* bucket_addr = (key_metas + bucketid * Associativity);
+            // find the hitted slot in hashnode
+            for (int i = 0; i != Associativity; i++) {
+                if (bucket_addr[i].isEmpty()) {
+                    return std::pair<bool, iterator>(
+                        false, iterator(false, T(), this, bucketid, i));
                 }
 
-                // update the targetNode
-                targetNode = burst(elements, key, keysize, this->anode::parent,
-                                   move_pos, hm);
+                std::pair<bool, T> res =
+                    find_kv_in_pages(bucket_addr[i], key, keysize);
+                if (res.first) {
+                    return std::pair<bool, iterator>(
+                        true, iterator(true, res.second, this, bucketid, i));
+                }
+            }
+            return std::pair<bool, iterator>(
+                false, iterator(false, T(), this, bucketid, -1));
+        }
 
-                // TODO: if have onlyValue, it has to be moved to the created
-                // trie_node
-                delete this;
-
-                if (targetNode->haveValue != false) {
-                    if (sp != nullptr) {
-                        sp->hnode = targetNode;
-                        sp->abucket = nullptr;
-                        sp->pos = 0;
-                    }
-                    return targetNode->onlyValue;
+        iterator search_kv_in_hashnode(const CharT* key, size_t keysize) {
+            if (keysize == 0) {
+                if (haveValue) {
+                    return iterator(true, onlyValue, this, 0, 0);
+                } else {
+                    return iterator(false, T(), this, 0, 0);
                 }
             }
 
-            size_t hashval = myTrie::hashRelative::hash<CharT>(
-                                 key + move_pos, keysize - move_pos) %
-                             hn_bucket_num;
-            if (sp != nullptr) {
-                sp->hnode = targetNode;
-                return targetNode->kvs[hashval].access_kv_in_bucket(
-                    key + move_pos, keysize - move_pos, sp, hm,
-                    &(targetNode->element_count));
-            } else {
-                return targetNode->kvs[hashval].access_kv_in_bucket(
-                    key + move_pos, keysize - move_pos, nullptr, hm,
-                    &(targetNode->element_count));
+            size_t bucketId1 =
+                myTrie::hashRelative::hash(key, keysize, 1) % Bucket_num;
+            std::pair<bool, iterator> res1 =
+                find_in_bucket(bucketId1, key, keysize);
+
+            size_t bucketId2 =
+                myTrie::hashRelative::hash(key, keysize, 2) % Bucket_num;
+            std::pair<bool, iterator> res2 =
+                find_in_bucket(bucketId2, key, keysize);
+
+            // if found the existed target in bucket1 or bucket2, just return
+            // the iterator for being modified or read
+            if (res1.first) {
+                return res1.second;
+            } else if (res2.first) {
+                return res2.second;
             }
+
+            // if the code reach here it means the target doesn't exist
+            // we return the iterator with empty slot
+            if (res1.second.slotid != -1) {
+                return res1.second;
+            } else if (res2.second.slotid != -1) {
+                return res2.second;
+            }
+            // if two bucket are both full, we return the res1's iterator with
+            // slotid == -1, and let the
+            // 1. findMode: found==false
+            // 2. !findMode:found==false, slotid==-1, need to kick some slot
+            return res1.second;
         }
 
-        std::pair<std::string, T> getSubStr(std::pair<std::string, T>& element,
-                                            size_t sub_pos = 1) {
-            return std::pair<std::string, T>(element.first.substr(sub_pos),
-                                             element.second);
+        inline slot* get_slot_addr(size_t bucketid, size_t slotid) {
+            return &key_metas[bucketid * Associativity + slotid];
+        }
+
+        bool need_burst() const {
+            return elem_num >= Max_slot_num * Burst_ratio;
         }
 
         // To turn this(a hashnode) to n childs of trie_node linking their
-        // child(hashnode or trienode)
-        hash_node* burst(std::map<std::string, T>& elements, const CharT* key,
-                         size_t keysize, trie_node* p, size_t& move_pos,
-                         htrie_map* hm) {
-            hash_node* target_hashNode = nullptr;
-
-            std::map<CharT, std::map<std::string, T>> splitElements;
+        // hashnode
+        void burst(std::map<std::string, T>& elements, trie_node* p,
+                   htrie_map* hm) {
+            std::map<CharT, std::map<std::string, T>> preprocElements;
             for (auto it = elements.begin(); it != elements.end(); it++) {
-                splitElements[(it->first)[0]][it->first.substr(1)] = it->second;
+                preprocElements[(it->first)[0]][it->first.substr(1)] =
+                    it->second;
             }
 
-            for (auto it = splitElements.begin(); it != splitElements.end();
+            for (auto it = preprocElements.begin(); it != preprocElements.end();
                  it++) {
-                trie_node* cur_trie_node = new trie_node(it->first, nullptr);
-
                 if (p == nullptr) {
                     // bursting in a root hashnode
                     // the t_root is update to a empty trie_node
-                    hm->t_root = new trie_node('\0', nullptr);
-                    cur_trie_node->anode::setParent((trie_node*)hm->t_root);
-                    ((trie_node*)hm->t_root)->addChildTrieNode(cur_trie_node);
-                    p = (trie_node*)hm->t_root;
-                } else {
-                    // bursting in a normal hashnode
-                    cur_trie_node->anode::setParent(p);
-                    p->addChildTrieNode(cur_trie_node);
+                    p = new trie_node('\0', nullptr);
+                    hm->setRoot(p);
                 }
 
-                if (it->first == *(key)) {
-                    move_pos++;
-                }
+                trie_node* cur_trie_node = new trie_node(it->first, p);
+                p->addChildTrieNode(cur_trie_node);
 
                 std::map<std::string, T>& curKV = it->second;
-
-                if (splitElements.size() == 1) {
-                    return burst(curKV, key + 1, keysize - 1, cur_trie_node,
-                                 move_pos, hm);
+                if (preprocElements.size() == 1) {
+                    return burst(curKV, cur_trie_node, hm);
                 }
 
-                hash_node* hnode =
-                    new hash_node(hn_burst_threshold, hn_bucket_num);
-                hnode->anode::setParent(cur_trie_node);
+                hash_node* hnode = new hash_node(cur_trie_node);
                 cur_trie_node->setOnlyHashNode(hnode);
 
+                bool stop_insert_and_burst = false;
                 for (auto itt = curKV.begin(); itt != curKV.end(); itt++) {
                     std::string temp = itt->first;
 
                     if (temp.size() == 0) {
                         hnode->haveValue = true;
                         hnode->onlyValue = itt->second;
-                        if (*(CharT*)key == cur_trie_node->anode::myChar &&
-                            myTrie::hashRelative::keyEqual(temp.data(),
-                                                           temp.size(), key + 1,
-                                                           keysize - 1)) {
-                            target_hashNode = hnode;
-                        }
-                        SearchPoint new_sp(hnode, nullptr, 0);
-                        hm->v2k[itt->second] = new_sp;
+
+                        hm->set_v2k(itt->second, hnode, nullptr);
                         continue;
                     }
 
-                    size_t hashval = myTrie::hashRelative::hash<CharT>(
-                                         temp.data(), temp.size()) %
-                                     hn_bucket_num;
-                    SearchPoint new_sp;
-                    new_sp.hnode = hnode;
-                    // write the value to the entry
-                    hnode->kvs[hashval].access_kv_in_bucket(
-                        temp.data(), temp.size(), &new_sp, hm,
-                        &(hnode->element_count)) = itt->second;
-                    hm->v2k[itt->second] = new_sp;
-
-                    if (*(CharT*)key == cur_trie_node->anode::myChar &&
-                        myTrie::hashRelative::keyEqual(temp.data(), temp.size(),
-                                                       key + 1, keysize - 1)) {
-                        target_hashNode = hnode;
+                    iterator target_it =
+                        hnode->search_kv_in_hashnode(temp.data(), temp.size());
+                    std::pair<bool, T> res = target_it.insertKV(
+                        temp.data(), temp.size(), hm, itt->second);
+                    // if insert failed, it need burst
+                    if (res.first == false) {
+                        stop_insert_and_burst = true;
+                        break;
                     }
                 }
-            }
-            return target_hashNode;
-        }
-    };
-
-    class array_bucket {
-       public:
-        CharT* arr_buffer;
-        size_t buffer_size;
-        static const KeySizeT END_OF_BUCKET =
-            std::numeric_limits<KeySizeT>::max();
-
-        array_bucket() {
-            // inited array_bucket: |END_OF_BUCKET|
-            arr_buffer = (CharT*)std::malloc(sizeof(END_OF_BUCKET));
-            std::memcpy(arr_buffer, &END_OF_BUCKET, sizeof(END_OF_BUCKET));
-            buffer_size = sizeof(END_OF_BUCKET);
-        }
-
-        bool is_end_of_bucket(const CharT* buffer) {
-            return read_key_size(buffer) == END_OF_BUCKET;
-        }
-
-        size_t read_key_size(const CharT* buffer) {
-            KeySizeT key_size;
-            std::memcpy(&key_size, buffer, sizeof(KeySizeT));
-
-            return (size_t)key_size;
-        }
-
-        // if found, return (entry_pos, true)
-        // if not found, return (inserting_pos, false)
-        std::pair<size_t, bool> find_in_bucket(const CharT* key,
-                                               size_t keysize) {
-            CharT* buffer_ptr = arr_buffer;
-            size_t pos = 0;
-            while (!is_end_of_bucket(buffer_ptr)) {
-                size_t length = read_key_size(buffer_ptr);
-                CharT* cmp_buffer_ptr = buffer_ptr + sizeof(KeySizeT);
-                if (myTrie::hashRelative::keyEqual(cmp_buffer_ptr, length, key,
-                                                   keysize)) {
-                    return std::pair<size_t, bool>(pos, true);
+                if (stop_insert_and_burst) {
+                    burst(curKV, cur_trie_node, hm);
+                    delete hnode;
                 }
-                // move ptr to next header, skip keysize, string, value
-                buffer_ptr = buffer_ptr + sizeof(KeySizeT) + length + sizeof(T);
-                pos += sizeof(KeySizeT) + length + sizeof(T);
             }
-            return std::pair<size_t, bool>(pos, false);
+            return;
         }
 
-        size_t cal_arrbuffer_newsize(size_t keysize) {
-            // when add a new key we need extra |size|string|value|
-            size_t new_buffer_size = buffer_size;
-            new_buffer_size +=
-                sizeof(KeySizeT) + keysize * sizeof(CharT) + sizeof(T);
-            return new_buffer_size;
-        }
-
-        void resetBucketElement_v2k(CharT* arr_buffer,
-                                    htrie_map<CharT, T>* hm) {
-            CharT* buffer_ptr = arr_buffer;
-            size_t pos = 0;
-            while (!is_end_of_bucket(buffer_ptr)) {
-                size_t length = read_key_size(buffer_ptr);
-                T* value_pos = (T*)(buffer_ptr + sizeof(KeySizeT) + length);
-
-                (hm->v2k[*value_pos]).pos = buffer_ptr;
-
-                // move ptr to next header, skip keysize, string, value
-                buffer_ptr = buffer_ptr + sizeof(KeySizeT) + length + sizeof(T);
-            }
-        }
-
-        T& access_kv_in_bucket(const CharT* target, size_t keysize,
-                               typename htrie_map<CharT, T>::SearchPoint* sp,
-                               htrie_map<CharT, T>* hm, uint32_t* counter) {
-            std::pair<size_t, bool> found = find_in_bucket(target, keysize);
-
-            if (found.second) {
-                // found the target, return the value reference
-                if (sp != nullptr) {
-                    sp->abucket = this;
-                    sp->pos = found.first;
-                }
-                return get_val_ref(found.first);
-
-            } else {
-                // not found, the buffer is full, need realloc
-                size_t new_size = cal_arrbuffer_newsize(keysize);
-
-                CharT* new_buffer =
-                    (CharT*)(std::realloc(arr_buffer, new_size));
-                if (new_buffer == nullptr) {
-                    std::cerr << "realloc failed!!!!!!!!!1" << std::endl;
-                    exit(-1);
-                }
-                arr_buffer = new_buffer;
-                buffer_size = new_size;
-
-                T v = T{};
-                append_impl(target, keysize, arr_buffer + found.first, v);
-                if (sp != nullptr) {
-                    sp->abucket = this;
-                    sp->pos = found.first;
-                }
-                (*counter)++;
-                return get_val_ref(found.first);
-            }
-        }
-
-        void append_impl(const CharT* target, size_t keysize,
+        void append_impl(const CharT* key, size_t keysize,
                          CharT* buffer_append_pos, T& value) {
-            // append key_size
-            std::memcpy(buffer_append_pos, &keysize, sizeof(KeySizeT));
-            buffer_append_pos += sizeof(KeySizeT);
-
             // append the string
-            std::memcpy(buffer_append_pos, target, keysize * sizeof(CharT));
+            std::memcpy(buffer_append_pos, key, keysize * sizeof(CharT));
             buffer_append_pos += keysize;
 
             // append the value
             std::memcpy(buffer_append_pos, &value, sizeof(T));
-            buffer_append_pos += sizeof(T);
-
-            // append the whole buffer end
-            const auto end_of_bucket = END_OF_BUCKET;
-            std::memcpy(buffer_append_pos, &end_of_bucket,
-                        sizeof(end_of_bucket));
         }
 
-        T& get_val_ref(size_t val_pos) {
-            CharT* entry_start_pos = arr_buffer + val_pos;
-            size_t length = read_key_size(entry_start_pos);
-            CharT* valAddress = entry_start_pos + sizeof(KeySizeT) + length;
-            return *((T*)valAddress);
-        }
+        // return: page_id, pos
+        std::pair<size_t, size_t> alloc_insert_space(size_t keysize) {
+            std::pair<char*, size_t> res = pages[cur_page_id];
 
-        void get_item_in_array_bucket(
-            std::vector<std::pair<std::string, T>>& res) {
-            CharT* buf = arr_buffer;
-            while (!is_end_of_bucket(buf)) {
-                size_t length = read_key_size(buf);
-
-                char* temp = (char*)malloc(length + 1);
-                std::memcpy(temp, buf + sizeof(KeySizeT), length);
-                temp[length] = '\0';
-                std::string item(temp);
-                free(temp);
-
-                // move ptr to next header, skip keysize, string, value
-                buf = buf + sizeof(KeySizeT) + length;
-                T v = *((T*)buf);
-
-                res.push_back(std::pair<std::string, T>(item, v));
-                buf = buf + sizeof(T);
+            // if the cur_page is full, malloc a new page
+            if (res.second + keysize * sizeof(CharT) + sizeof(T) >
+                Max_bytes_per_kv) {
+                char* page = (char*)malloc(Max_bytes_per_kv);
+                // set up the page information
+                pages.push_back(std::pair<char*, size_t>(
+                    page, keysize * sizeof(CharT) + sizeof(T)));
+                cur_page_id++;
+                return std::pair<size_t, size_t>(cur_page_id, 0);
             }
+            size_t offset = res.second;
+            // update the page information
+            pages[cur_page_id].second += keysize * sizeof(CharT) + sizeof(T);
+            return std::pair<size_t, size_t>(cur_page_id, offset);
+        }
+
+        void get_tail_str_v(std::map<std::string, T>& elements, slot* s) {
+            char* tail_pointer = get_tail_pointer(s);
+            std::string res(tail_pointer, s->length);
+            std::memcpy(&elements[res], tail_pointer + s->length, sizeof(T));
+        }
+
+        T get_tail_v(slot* s) {
+            T v;
+            std::memcpy(&v, get_tail_pointer(s) + s->length, sizeof(T));
+            return v;
+        }
+
+        void get_all_elements(std::map<std::string, T>& elements) {
+            for (size_t i = 0; i != Bucket_num; i++) {
+                for (size_t j = 0; j != Associativity; j++) {
+                    slot& cur_slot = key_metas[i * Associativity + j];
+                    if (cur_slot.isEmpty()) break;
+                    get_tail_str_v(elements, &cur_slot);
+                }
+            }
+        }
+
+        // return another possible bucketid that the slot *s can be
+        inline size_t get_another_bucketid(slot* s, size_t current_bucketid) {
+            size_t bucketid1 =
+                myTrie::hashRelative::hash(get_tail_pointer(s), s->length, 1) %
+                Bucket_num;
+            size_t bucketid2 =
+                myTrie::hashRelative::hash(get_tail_pointer(s), s->length, 2) %
+                Bucket_num;
+            return current_bucketid == bucketid1 ? bucketid2 : bucketid1;
+        }
+
+        inline slot* get_bucket(size_t bucketid) {
+            return key_metas + bucketid * Associativity;
+        }
+
+        inline slot* get_slot(size_t bucketid, size_t slotid) {
+            return key_metas + bucketid * Associativity + slotid;
+        }
+
+        inline slot* previous_dst_slot_in_same_bucket(slot* s) {
+            size_t slotid = (s - key_metas) % Associativity;
+            if (slotid == 0) {
+                return nullptr;
+            } else
+                return s - 1;
+        }
+
+        void apply_the_changed_searchPoint(map<T, slot*>& searchPoints,
+                                           htrie_map<CharT, T>* hm) {
+            for (auto it = searchPoints.begin(); it != searchPoints.end(); it++)
+                hm->set_searchPoint_slot(it->first, it->second);
+        }
+
+        int rehash(size_t bucketid, htrie_map<CharT, T>* hm) {
+            char* key_metas_backup =
+                (char*)malloc(Bucket_num * Associativity * sizeof(slot));
+            memcpy(key_metas_backup, key_metas,
+                   Bucket_num * Associativity * sizeof(slot));
+
+            // bucket_list records the mapping of bucket_id=last_empty_slot_id
+            std::map<size_t, size_t> bucket_list;
+            for (size_t bn = 0; bn != Bucket_num; bn++) {
+                bucket_list[bn] = Associativity;
+                for (int sn = 0; sn != Associativity; sn++) {
+                    if (key_metas[bn * Associativity + sn].isEmpty()) {
+                        bucket_list[bn] = sn;
+                        break;
+                    }
+                }
+            }
+            // current bucket is definitely full
+            // just pick the last slot to kick
+            int ret_slot_id = Associativity - 1;
+
+            size_t kicked_slot_id = -1;
+            for (int i = 0; i != Associativity; i++) {
+                slot* s = get_slot(bucketid, i);
+                size_t bkid = get_another_bucketid(s, bucketid);
+                if (bkid != bucketid) {
+                    kicked_slot_id = i;
+                }
+            }
+            if (kicked_slot_id == -1) {
+                return -1;
+            }
+            ret_slot_id = kicked_slot_id;
+
+            size_t current_bucket_id = bucketid;
+
+            slot src_slot = slot(0, 0, 0);
+            slot* dst_slot = get_slot(current_bucket_id, kicked_slot_id);
+
+            size_t rehash_count = 0;
+
+            size_t last_current_bucketid = 0;
+            size_t last_bucketid_kick_to = 0;
+
+            map<T, slot*> searchPoint_wait_2_be_update;
+            do {
+                /*
+                    src(a,b,c)
+                    cur_bucket: |x      |x      |x      |dst(d,e,f)|
+                    kk2_bucket: |x      |x      |x      |x         |
+                */
+                // calculate the destination
+                size_t bucketid_kick_to =
+                    get_another_bucketid(dst_slot, current_bucket_id);
+
+                // if the slot can only place in one bucket, we change the
+                // dst_slot
+                // if the cuckoo hash kick as a circle, we change the dst_slot
+                // to try to break the circle
+                if (bucketid_kick_to == current_bucket_id ||
+                    (last_bucketid_kick_to == current_bucket_id &&
+                     last_current_bucketid == bucketid_kick_to)) {
+                    dst_slot = previous_dst_slot_in_same_bucket(dst_slot);
+                    rehash_count++;
+                    if (dst_slot == nullptr) {
+                        // recover the key_metas
+                        memcpy(key_metas, key_metas_backup,
+                               Bucket_num * Associativity * sizeof(slot));
+                        return -1;
+                    }
+                    continue;
+                }
+
+                /*
+                    src(a,b,c)
+                    temp: d,e,f
+                    cur_bucket: |x      |x      |x      |dst(d,e,f)|
+                    kk2_bucket: |x      |x      |x      |x         |
+                */
+                KeySizeT temp_length = dst_slot->length;
+                size_t temp_pos = dst_slot->pos;
+                size_t temp_page_id = dst_slot->page_id;
+
+                // if dst_slot is empty, it means now the dst_slot is the first
+                // place we clear for the target slot
+                if (dst_slot->isEmpty()) {
+                    // recover the key_metas
+                    memcpy(key_metas, key_metas_backup,
+                           Bucket_num * Associativity * sizeof(slot));
+                    return -1;
+                }
+
+                /*
+                    src(a,b,c)
+                    temp: d,e,f
+                    cur_bucket: |x      |x      |x      |dst(a,b,c)|
+                    kk2_bucket: |x      |x      |x      |x         |
+                */
+                dst_slot->set_slot(src_slot.length, src_slot.pos,
+                                   src_slot.page_id);
+                if (!dst_slot->isEmpty())
+                    searchPoint_wait_2_be_update[get_tail_v(dst_slot)] =
+                        dst_slot;
+
+                // if the destination bucket isn't full, just fill the empty
+                // slot and return
+                if (bucket_list[bucketid_kick_to] != Associativity) {
+                    /* kick2bucket is have empty slot
+                        src(a,b,c)
+                        temp: d,e,f
+                        cur_bucket: |x      |x      |x      |dst(a,b,c)|
+                        kk2_bucket: |x      |x      |x      |0         |
+                    */
+                    // update dst_slot to the dest bucket's first empty slot
+                    /*
+                        src(a,b,c)
+                        temp: d,e,f
+                        cur_bucket: |x      |x      |x      |x(a,b,c)  |
+                        kk2_bucket: |x      |x      |x      |0-dst     |
+                    */
+                    dst_slot = get_slot(bucketid_kick_to,
+                                        bucket_list[bucketid_kick_to]);
+
+                    /*
+                        src(a,b,c)
+                        temp: d,e,f
+                        cur_bucket: |x      |x      |x      |x(a,b,c)  |
+                        kk2_bucket: |x      |x      |x      |dst(d,e,f)|
+                    */
+                    dst_slot->set_slot(temp_length, temp_pos, temp_page_id);
+                    searchPoint_wait_2_be_update[get_tail_v(dst_slot)] =
+                        dst_slot;
+                    apply_the_changed_searchPoint(searchPoint_wait_2_be_update,
+                                                  hm);
+                    return ret_slot_id;
+                }
+
+                /* kick2bucket is full
+                    src(a,b,c)
+                    temp: d,e,f
+                    cur_bucket: |x      |x      |x      |dst(a,b,c)|
+                    kk2_bucket: |x      |x      |x      |x         |
+                */
+
+                /* kick2bucket is full
+                    src(a,b,c)
+                    temp: d,e,f
+                    cur_bucket: |x      |x      |x      |x(a,b,c)  |
+                    kk2_bucket: |x      |x      |x      |x-dst     |
+                */
+                // update dst_slot to the dest bucket's last empty slot
+                dst_slot = get_slot(bucketid_kick_to, Associativity - 1);
+                /*
+                     src(d,e,f)
+                     cur_bucket: |x      |x      |x      |x(a,b,c)  |
+                     kk2_bucket: |x      |x      |x      |x-dst     |
+                */
+                src_slot.set_slot(temp_length, temp_pos, temp_page_id);
+                /*
+                     src(d,e,f)
+                     (kk2_bucket)
+                     cur_bucket: |x      |x      |x      |x-dst     |
+                */
+                last_bucketid_kick_to = bucketid_kick_to;
+                last_current_bucketid = current_bucket_id;
+                current_bucket_id = bucketid_kick_to;
+
+                rehash_count++;
+            } while (rehash_count != Max_loop);
+            // recover the key_metas
+            memcpy(key_metas, key_metas_backup,
+                   Bucket_num * Associativity * sizeof(slot));
+            return -1;
+        }
+
+        std::pair<bool, T> insert_kv_in_hashnode(const CharT* key,
+                                                 size_t keysize, htrie_map* hm,
+                                                 T v, size_t bucketid,
+                                                 int slotid) {
+            if (keysize == 0) {
+                haveValue = true;
+                onlyValue = v;
+                hm->set_v2k(v, this, nullptr);
+                return std::pair<bool, T>(true, v);
+            }
+
+            // if slotid==-1, it denotes that the bucket(bucketid) is full , so
+            // we rehash the key_metas
+            if (slotid == -1) {
+                if ((slotid = rehash(bucketid, hm)) == -1) {
+                    return std::pair<bool, T>(false, T());
+                }
+            }
+
+            // now the slotid cannot be -1 and slotid is lower than
+            // Associativity
+            assert(slotid != -1 && slotid >= 0 && slotid < Associativity);
+
+            slot* target_slot =
+                key_metas + Associativity * bucketid + (size_t)slotid;
+
+            // allocate new page or alloc more space in old page
+            std::pair<size_t, size_t> res = alloc_insert_space(keysize);
+
+            target_slot->length = keysize;
+            target_slot->page_id = res.first;
+            target_slot->pos = res.second;
+            append_impl(key, keysize, get_tail_pointer(target_slot), v);
+
+            // set v2k
+            hm->set_v2k(v, this, target_slot);
+            elem_num++;
+
+            // todo: need to burst elegantly
+            if (need_burst()) {
+                std::map<std::string, T> elements;
+                get_all_elements(elements);
+
+                burst(elements, this->anode::parent, hm);
+
+                delete this;
+            }
+
+            return std::pair<bool, T>(true, v);
         }
     };
 
     class SearchPoint {
        public:
         hash_node* hnode;
-        array_bucket* abucket;
-        uint32_t pos;
+        typename hash_node::slot* sl;
 
-        SearchPoint() : hnode(nullptr), pos(0) {}
-        SearchPoint(hash_node* h, array_bucket* ab, uint32_t p)
-            : hnode(h), abucket(ab), pos(p) {}
+        SearchPoint() : hnode(nullptr), sl(nullptr) {}
+        SearchPoint(hash_node* h, typename hash_node::slot* s)
+            : hnode(h), sl(s) {}
 
+        void set_slot(typename hash_node::slot* s) { sl = s; }
+
+        // todo: refine
         std::string getString() {
+            // get the parent char chain
             trie_node* cur_node = hnode->anode::parent;
             std::stack<char> st;
-            while (cur_node != nullptr && cur_node->anode::myChar != '\0') {
-                st.push((char)cur_node->anode::myChar);
+            while (cur_node != nullptr && cur_node->myChar != '\0') {
+                st.push((char)cur_node->myChar);
                 cur_node = cur_node->anode::parent;
             }
             std::stringstream ss;
@@ -477,19 +653,10 @@ class htrie_map {
                 ss << st.top();
                 st.pop();
             }
-            if (abucket != nullptr) {
-                std::string res;
-                KeySizeT key_size;
-                CharT* buf = abucket->arr_buffer;
-                std::memcpy(&key_size, buf + pos, sizeof(KeySizeT));
-
-                size_t length = (size_t)key_size;
-
-                char* temp = (char*)malloc(length + 1);
-                std::memcpy(temp, buf + pos + sizeof(KeySizeT), length);
-                temp[length] = '\0';
-                res = std::string(temp);
-                free(temp);
+            // get tail
+            if (sl != nullptr) {
+                string res = std::string(hnode->get_tail_pointer(sl),
+                                         (size_t)sl->length);
                 ss << res;
             }
             return ss.str();
@@ -499,163 +666,115 @@ class htrie_map {
    public:
     anode* t_root;
     std::map<T, SearchPoint> v2k;
-    htrie_map(size_t customized_burst_threshold = DEFAULT_BURST_THRESHOLD,
-              size_t customized_bucket_count = DEFAULT_BURST_THRESHOLD) {
-        t_root =
-            new hash_node(customized_burst_threshold, customized_bucket_count);
-        burst_threshold = customized_burst_threshold;
-        bucket_num = customized_bucket_count;
+    htrie_map(size_t customized_associativity = DEFAULT_Associativity,
+              size_t customized_bucket_count = DEFAULT_Bucket_num,
+              size_t customized_byte_per_kv = DEFAULT_Max_bytes_per_kv,
+              double customized_burst_ratio = DEFAULT_Burst_ratio)
+        : t_root(nullptr) {
+        Associativity = customized_associativity;
+        Bucket_num = customized_bucket_count;
+        Max_bytes_per_kv = customized_byte_per_kv;
+        Burst_ratio = customized_burst_ratio;
+
+        Max_slot_num = Associativity * Bucket_num;
+        Max_loop = Max_slot_num * 0.5;
+
+        t_root = new hash_node(nullptr);
+    }
+
+    void set_searchPoint_slot(T v, typename hash_node::slot* s) {
+        v2k[v].set_slot(s);
+    }
+
+    void set_v2k(T v, hash_node* hnode, typename hash_node::slot* s) {
+        v2k[v] = SearchPoint(hnode, s);
     }
 
     void setRoot(anode* node) { t_root = node; }
 
-    const T searchByKey(std::string key) {
-        return find(key.data(), key.size(), nullptr);
+    std::pair<bool, T> searchByKey(std::string key) {
+        return access_kv_in_htrie_map(key.data(), key.size(), T(), true);
     }
 
     std::string searchByValue(T v) { return v2k[v].getString(); }
 
-    bool insertKV(std::string key, T v) {
-        SearchPoint sp;
-        find(key.data(), key.size(), &sp) = v;
-        v2k[v] = sp;
-        return true;
+    std::pair<bool, T> insertKV(std::string key, T v) {
+        return access_kv_in_htrie_map(key.data(), key.size(), v, false);
     }
 
-    T& find(const CharT* key, size_t key_size, SearchPoint* sp) {
-        anode* current_node = t_root;
-        if (t_root->isHashNode()) {
-            // find a key in hash_node
-            // if find: return the v reference
-            // if notfind: write the key and return the v reference
-            return ((hash_node*)current_node)
-                ->access_kv_in_hashnode(key, key_size, this, sp);
+    class iterator {
+       public:
+        bool found;
+        const T v;
+        hash_node* target_node;
+        size_t bucketid;
+        int slotid;
+
+        iterator(bool f, T vv, hash_node* hnode, size_t bid, int sid)
+            : found(f), v(vv), target_node(hnode), bucketid(bid), slotid(sid) {}
+
+        std::pair<bool, T> insertKV(const CharT* key, size_t key_size,
+                                    htrie_map<CharT, T>* hm, T v) {
+            return target_node->insert_kv_in_hashnode(key, key_size, hm, v,
+                                                      bucketid, slotid);
         }
+    };
+
+    std::pair<bool, T> access_kv_in_htrie_map(const CharT* key, size_t key_size,
+                                              T v, bool findMode) {
+        anode* current_node = t_root;
 
         for (size_t pos = 0; pos < key_size; pos++) {
             if (current_node->isTrieNode()) {
-                anode* parent = current_node;
-                current_node =
-                    ((trie_node*)current_node)->findChildNode(key[pos]);
+                trie_node* parent;
+                parent = (trie_node*)current_node;
 
-                if (current_node == nullptr) {
-                    // can't find, create a relative trie_node and a hashnode
-                    // child, set current_node as the trie_node
-                    current_node = ((trie_node*)parent)
-                                       ->create_trienode_With(
-                                           key[pos], this->burst_threshold,
-                                           this->bucket_num);
-                } else if (current_node->isHashNode()) {
-                    // if find the target hashnode instead of moving the pos,
-                    // pos should be recover
-                    pos--;
-                }
+                // only return the hitted trie_node* or nullptr if not found
+                current_node = parent->findChildNode(key[pos], findMode);
+
             } else {
-                return ((hash_node*)current_node)
-                    ->access_kv_in_hashnode(key + pos, key_size - pos, this,
-                                            sp);
+                iterator it =
+                    ((hash_node*)current_node)
+                        ->search_kv_in_hashnode(key + pos, key_size - pos);
+                if (findMode) {
+                    return std::pair<bool, T>(it.found, it.v);
+                } else {
+                    pair<bool, T> res =
+                        it.insertKV(key + pos, key_size - pos, this, v);
+                    if (res.first == false) {
+                        // if the insert failed, we burst the target_hashnode
+                        // and retry insertion
+                        hash_node* hnode_burst_needed =
+                            (hash_node*)current_node;
+                        map<string, T> hnode_elems;
+                        hnode_burst_needed->get_all_elements(hnode_elems);
+                        hnode_burst_needed->burst(
+                            hnode_elems, hnode_burst_needed->anode::parent,
+                            this);
+                        delete hnode_burst_needed;
+                        return access_kv_in_htrie_map(key, key_size, v, false);
+                    }
+                    return res;
+                }
+            }
+            // only in the findMode==true can cause the current_node to be
+            // nullptr
+            if (current_node == nullptr) {
+                return std::pair<bool, T>(false, T());
             }
         }
-        return ((trie_node*)current_node)
-            ->getOnlyHashNode()
-            ->access_onlyValue_in_hashnode(this, sp);
+
+        // find a key in hash_node's only value
+        iterator it = ((hash_node*)current_node)->search_kv_in_hashnode(key, 0);
+
+        if (findMode) {
+            return std::pair<bool, T>(it.found, it.v);
+        } else {
+            return it.insertKV(key, 0, this, v);
+        }
     }
 
     void deleteMyself() { t_root->deleteMe(); }
 };  // namespace myTrie
 
-}  // namespace myTrie
-
-namespace myTrie {
-namespace debuging {
-using std::cout;
-using std::endl;
-uint64_t h_n;
-uint64_t t_n;
-uint64_t hash_node_mem;
-uint64_t trie_node_mem;
-
-template <typename CharT, typename T>
-void print_tree_construct(typename myTrie::htrie_map<CharT, T>::anode* root) {
-    using my = typename myTrie::htrie_map<CharT, T>;
-    if (root == nullptr) {
-        return;
-    }
-    // print bucket
-    if (root->isHashNode()) {
-        hash_node_mem += sizeof(root);
-
-        std::vector<typename my::array_bucket> buckets =
-            ((typename my::hash_node*)root)->kvs;
-        // basic bucket cost
-        uint32_t bucket_num = buckets.size();
-        uint32_t bucket_mem = sizeof(typename my::array_bucket) * bucket_num;
-        hash_node_mem += bucket_mem;
-
-        uint32_t bucket_buffer_mem = 0;
-        // bucket buffer cost
-        for (auto it = buckets.begin(); it != buckets.end(); it++) {
-            bucket_buffer_mem += it->buffer_size;
-        }
-        hash_node_mem += bucket_buffer_mem;
-        h_n++;
-
-    } else if (root->isTrieNode()) {
-        trie_node_mem += sizeof(root);
-
-        std::map<CharT, typename myTrie::htrie_map<CharT, T>::anode*> childs =
-            ((typename myTrie::htrie_map<CharT, T>::trie_node*)root)->childs;
-        if (childs.size() == 0) {
-            print_tree_construct<CharT, T>(
-                ((typename myTrie::htrie_map<CharT, T>::trie_node*)root)
-                    ->getOnlyHashNode());
-        } else {
-            for (auto it = childs.begin(); it != childs.end(); it++) {
-                t_n++;
-                print_tree_construct<CharT, T>(it->second);
-            }
-        }
-    } else {
-        std::cout << "node is not trie nor hash node\n";
-        exit(0);
-    }
-    return;
-}
-
-template <typename CharT, typename T>
-double print_res() {
-    std::cout << "trie_node: " << t_n << " hash_node: " << h_n << std::endl;
-    std::cout << "trie_node_mem: " << trie_node_mem
-              << " hash_node_mem: " << hash_node_mem << std::endl;
-    using my = typename myTrie::htrie_map<CharT, T>;
-
-    std::cout << "total: ," << (hash_node_mem + trie_node_mem) / 1024 / 1024
-              << ", mb" << std::endl;
-    double ret_v;
-    ret_v = (hash_node_mem + trie_node_mem) / 1024 / 1024;
-
-    // todo
-    // std::cout << "trienode size:" << sizeof(typename my::trie_node)
-    //           << std::endl;
-    // std::cout << "hashnode size:" << sizeof(typename my::hash_node)
-    //           << std::endl;
-    // std::cout << "arraybucket size:" << sizeof(typename my::array_bucket)
-    //           << std::endl;
-    // std::cout << "htrie_map size:"
-    //           << sizeof(typename myTrie::htrie_map<CharT, T>) << std::endl;
-    // std::cout
-    //     << "htrie_map::v2k size:"
-    //     << sizeof(
-    //            std::map<T, typename myTrie::htrie_map<CharT,
-    //            T>::SearchPoint>)
-    //     << std::endl;
-    // std::cout << "arraybucket::kvs size:"
-    //           << sizeof(std::vector<typename my::array_bucket>) << std::endl;
-    t_n = 0;
-    h_n = 0;
-    hash_node_mem = 0;
-    trie_node_mem = 0;
-    return ret_v;
-}
-}  // namespace debuging
 }  // namespace myTrie
